// models/funcionario.model.ts
type FuncionarioModel = {
  codigo?: number;
  agencia?: number;
  codigoBanco?: number;
  digitoConta?: string;
  digitoAgencia?: string;
  numeroConta?: number;
  categoriaCnh?: string;
  cnh?: string;
  cpf?: string;
  ctps?: string;
  ctpsUf?: string;
  dataAdmissao?: string;
  dataEmissaoCnh?: string;
  dataEmissaoRg?: string;
  dataNascimento?: string;
  dataValidadeCnh?: string;
  nomeEmpresa?: string;
  nomeMunicipio?: string;
  codigoEmpresa?: number;
  codigoEscritorioAreaEmpresas?: number;
  logradouro?: string;
  numero?: string;
  complemento?: string;
  bairro?: string;
  codigoUf?: number;
  codigoMunicipio?: number;
  cep?: string;
  siglaUF?: string;
  deficiencia?: string;
  deficienciaAuditiva?: boolean;
  deficienciaFisica?: boolean;
  deficienciaIntelectual?: boolean;
  deficienciaPsicossocial?: boolean;
  deficienciaVisual?: boolean;
  email?: string;
  estadoCivil?: number;
  identidadeGenero?: number;
  nacionalidade?: string;
  naturalidade?: string;
  nome?: string;
  nomeMae?: string;
  nomePai?: string;
  nomeSocial?: string;
  numeroDependentes?: number;
  numeroFilhos?: number;
  numeroReservista?: number;
  orgaoEmissorRg?: string;
  pisPasep?: string;
  rg?: string;
  secaoTituloEleitor?: number;
  serieCTPS?: string;
  tipoRaca?: number;
  tituloEleitor?: string;
  zonaTituloEleitor?: number;
  matricula?: string;
  ativo?: boolean;
  motivoDesligamento?: number;
  dataDesligamento?: string;
  vinculo?: string;
  perfil?: string;
  codigoLider?: number;
  codigoLiderAprovacao?: number;
  codigoCoach?: number;
  codigoArea?: number;
  codigoCargo?: number;
  codigoFuncao?: number;
  apelido?: string;
  nomeArea?: string;
  nomeCargo?: string;
  nomeFuncao?: string;
  nomeLider?: string;
  nomeCoach?: string;
  isChecked?: boolean;
};

export type { FuncionarioModel };